+++ src/fr/maxlego08/menu/api/button/PerformButton.java
@@ -1,6 +1,7 @@
 package fr.maxlego08.menu.api.button;
 
 import fr.maxlego08.menu.api.scheduler.ZScheduler;
+import fr.maxlego08.menu.api.utils.Placeholders;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.ClickType;
 
@@ -58,8 +59,8 @@ public interface PerformButton {
      *
      * @param player    The player who will execute the commands.
      * @param type      The click type.
-     * @param scheduler
+     * @param scheduler The ZScheduler used to schedule the command executions.
      */
-    void execute(Player player, ClickType type, ZScheduler scheduler);
+    void execute(Player player, ClickType type, ZScheduler scheduler, Placeholders placeholders);
 
 }